semantic projection works!

+++ fuel_planner/plan_env/include/plan_env/sdf_map.h
@@ -115,7 +115,7 @@ struct MapData {
   std::vector<double> distance_buffer_;
   std::vector<double> tmp_buffer1_;
   std::vector<double> tmp_buffer2_;
-  std::vector<uint8_t> semantics_buffer_;
+  std::vector<uint32_t> semantics_buffer_;
   // data for updating
   vector<short> count_hit_, count_miss_, count_hit_and_miss_;
   vector<char> flag_rayend_, flag_visited_;
